simplify component file structure
